refactor: move functions in `sort-union-or-intersection-types` common folder
