+++ rules/sort-union-or-intersection-types/compute-matched-context-options.ts
