+++ rules/sort-union-or-intersection-types/compute-node-name.ts
