+++ rules/sort-union-or-intersection-types/type-contains-callable-type.ts
